pyproject.toml
[tool.poetry]
name = "party_wall_django"
version = "0.1.0"
description = "Party Wall application backend implemented with Django + DRF"
authors = ["15bubbles <15bubbles.15bubbles@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2 = "^2.8.6"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
gunicorn = "^20.0.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
mypy = "^0.790"
isort = "^5.7.0"
flake8 = "^3.8.4"
autoflake = "^1.4"
pytest = "^6.2.1"
pytest-django = "^4.1.0"
pytest-mock = "^3.5.1"
pytest-black = "^0.3.12"
pytest-pylint = "^0.18.0"
pytest-cov = "^2.11.1"
pytest-randomly = "^3.5.0"
pytest-sugar = "^0.9.4"
pylint-django = "^2.4.2"

[tool.black]
line-length = 110
target-version = ["py38"]

[tool.isort]
multi_line_output = 3
line_length = 110
include_trailing_comma = true

[tool.pylint]
init-hook = "import sys; sys.path.append('.')"
load-plugins = "pylint-django"

[tool.pylint.messages_control]
max-line-length = 110
ignore = "models"
disable = "missing-module-docstring,missing-class-docstring,missing-function-docstring,raise-missing-from,redefined-builtin"
good-names="q,healthcheck,id"

[tool.pytest.ini_options]
minversion = "6.0"
required_plugins = "pytest-mock"
addopts = "--color=yes --code-highlight=yes"
DJANGO_SETTINGS_MODULE = "party_wall_django.settings.test"
python_files = "tests.py test_*.py *_tests.py"

[tool.coverage.run]
source = "."
omit = "tests"

[tool.coverage.report]
skip_empty = true

[tool.coverage.xml]
output = "coverage_report.xml"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
